Move Entry.h and EntryType.h to schema/Entry/ folder, updated include paths in schema/
